requirements.txt
gevent
requests
Flask
Flask-SQLAlchemy
Flask-Script
Flask-Testing
Flask-Cors
Flask-Migrate
flask-socketio
PyJWT
validators
SQLAlchemy
paho-mqtt
